Modify circular buffer code for pre and post recording
changes to allow to pre record into a circular buffer and when hitting a key
or getting a signal the video is written directly to disk, with a offset in
the file to fill in the circular buffer later, after the recording has
finished the ring buffer is written to the beginning of the recorded file.

Example: raspivid -k -c 5000 -t 5000 -o out.h264
This will create a 10 sec video starting from 5 sec before hitting the key.