Add option to copy h264 stream to stdout
Option --stdout or -cp will cop the h264 encoded video to stdout.
i.e., raspivid -o test.mp4 -cp | nc $IP 5001
Can be used together with the circular buffer function to monitor stream
over netcat and sending signal to save buffer. In this mode all encoded
buffer will be sent.

For the "keypress,signal and intervall modes" it will not send data to
stdout if the toggle state->bCapturing is set to "pause". Then no encoding
is done and nothing can be send.